Some error handling fixes. API best practices for naming

+++ library/src/main/java/com/hpursan/digicert/library/service/BookService.java
@@ -34,6 +34,9 @@ public class BookService {
     }
 
     public void deleteBook(Long id){
+        if (!bookRepository.existsById(id)) {
+            throw new BookNotFoundException("Book with id " + id + " not found");
+        }
         bookRepository.deleteById(id);
     }
